LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/path-objects/2d.path.isPointInPath.empty.worker-expected.txt


PASS isPointInPath() works when there is no path
